pyproject.toml
[project]
name = "volley"
version = "0.1.0"
description = "RL agents for 3d pong. Game engine included in set."
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "msgpack>=1.1.0",
    "pyzmq>=27.1.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "stable-baselines3>=2.2.0",
    "tensorboard>=2.15.0",
    "torch>=2.1.0",
]
